cleanup poackage structure
